requirements.txt
faster-whisper==1.2.0
ffmpeg-python==0.2.0
uvicorn==0.35.0
python-multipart==0.0.20
aiogram==3.22.0
fastapi==0.116.1
python-dotenv==1.1.1